# docs/growthpilot/api/openapi.yaml
openapi: 3.1.0
info:
  title: 洪基托管成长中心后台系统 API
  version: v1.0.0
  description: 移除法律合规和心理筛查后的核心后台 API 设计稿。
servers:
- url: https://api.hongji-growth.local/api/v1
  description: Design Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    pageNo:
      name: pageNo
      in: query
      schema:
        type: integer
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
    keyword:
      name: keyword
      in: query
      schema:
        type: string
...
    sortBy:
      name: sortBy
      in: query
      schema:
        type: string
    sortOrder:
      name: sortOrder
      in: query
      schema:
        type: string
        enum:
        - asc
        - desc
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data: {}
        traceId:
          type: string
    PageResultMeta:
      type: object
      properties:
        pageNo:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
security:
- bearerAuth: []
paths:
...
  /tasks:
    get:
      tags:
      - tasks
      summary: 任务列表
      parameters:
      - $ref: '#/components/parameters/keyword'
      - $ref: '#/components/parameters/pageNo'
      - $ref: '#/components/parameters/pageSize'
      - $ref: '#/components/parameters/sortBy'
      - $ref: '#/components/parameters/sortOrder'
      - name: ownerUserId
        in: query
        schema:
          type: string
          format: uuid
      - name: studentId
        in: query
        schema:
          type: string
          format: uuid
      - name: familyId
        in: query
        schema:
          type: string
          format: uuid
      - name: teacherId
        in: query
        schema:
          type: string
          format: uuid
      - name: taskType
        in: query
        schema:
          type: string
      - name: priority
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      - name: dateFrom
        in: query
        schema:
          type: string
          format: date
      - name: dateTo
        in: query
        schema:
          type: string
          format: date
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        list:
                          type: array
                          items:
                            type: object
                            additionalProperties: true
                        page:
                          $ref: '#/components/schemas/PageResultMeta'
    post:
      tags:
      - tasks
      summary: 创建任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - taskType
              - ownerUserId
              - title
              properties:
                taskType:
                  type: string
                ownerUserId:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                dueAt:
                  type: string
                  format: date-time
                sourceType:
                  type: string
                sourceId:
                  type: string
                  format: uuid
                studentId:
                  type: string
                  format: uuid
                familyId:
                  type: string
                  format: uuid
                teacherId:
                  type: string
                  format: uuid
                resultNote:
                  type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      additionalProperties: true
  /tasks/{taskId}:
    patch:
      tags:
      - tasks
      summary: 更新任务状态
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                resultNote:
                  type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      additionalProperties: true
  /alerts:
    get:
      tags:
      - alerts
      summary: 预警列表
      parameters:
      - $ref: '#/components/parameters/keyword'
      - $ref: '#/components/parameters/pageNo'
      - $ref: '#/components/parameters/pageSize'
      - $ref: '#/components/parameters/sortBy'
      - $ref: '#/components/parameters/sortOrder'
      - name: studentId
        in: query
        schema:
          type: string
          format: uuid
      - name: familyId
        in: query
        schema:
          type: string
          format: uuid
      - name: invoiceId
        in: query
        schema:
          type: string
          format: uuid
      - name: alertType
        in: query
        schema:
          type: string
      - name: alertLevel
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      - name: dateFrom
        in: query
        schema:
          type: string
          format: date
      - name: dateTo
        in: query
        schema:
          type: string
          format: date
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        list:
                          type: array
                          items:
                            type: object
                            additionalProperties: true
                        page:
                          $ref: '#/components/schemas/PageResultMeta'
    post:
      tags:
      - alerts
      summary: 创建预警
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - alertType
              - title
              - content
              properties:
                alertType:
                  type: string
                alertLevel:
                  type: string
                title:
                  type: string
                content:
                  type: string
                sourceType:
                  type: string
                sourceId:
                  type: string
                  format: uuid
                studentId:
                  type: string
                  format: uuid
                familyId:
                  type: string
                  format: uuid
                invoiceId:
                  type: string
                  format: uuid
                resolverUserId:
                  type: string
                  format: uuid
                resolvedAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      additionalProperties: true
  /alerts/{alertId}:
    patch:
      tags:
      - alerts
      summary: 更新预警状态
      parameters:
      - name: alertId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                resolverUserId:
                  type: string
                  format: uuid
                resolvedAt:
                  type: string
                  format: date-time
                content:
                  type: string
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      additionalProperties: true
...